Add output file naming option and improve data handling
A new command-line argument "--out_suffix" is added to allow for specified output file names. This will prevent overwriting of previously produced data if the script is run multiple times. Also, data handling has been enhanced by encoding augments conditionally (based on whether it is the first run of the process) and improving the consolidation of test sequences.