Add argument to use last logits and enhance evaluation
An argument was added to the training script to optionally use the last logits for the calculation of the loss function. The evaluation script was also upgraded with extended sequence processing and the inclusion of a confusion matrix. Additionally, the new class_names dictionary was added to prepare_plasticc.py for easier class identification.