Update training and evaluation scripts
The training script now includes an option for label smoothing in the loss functions. The option name for using last labels only has also been updated for clarity. Additionally, the evaluation script has been enhanced to include an output directory for saving a confusion matrix and a limit on the number of sequences to be tested. Random shuffling of test sequences is also now included for randomization.